fdo_agi_repo/requirements_rpa.txt
# ============================================================
# Phase 2.5 RPA + YouTube Learning - Python Dependencies
# ============================================================
# 설치: pip install -r requirements_rpa.txt
# ============================================================

# ============================================================
# RPA Core Infrastructure
# ============================================================
pyautogui>=0.9.54          # 마우스/키보드 제어
pywinauto>=0.6.8           # Windows GUI 자동화
pynput>=1.7.6              # 키보드/마우스 이벤트 감지

# ============================================================
# OCR & Computer Vision
# ============================================================
pytesseract>=0.3.10        # Tesseract OCR Python 래퍼
easyocr>=1.7.0             # GPU 지원 OCR (더 정확)
opencv-python>=4.8.0       # 이미지/영상 처리
pillow>=10.1.0             # 이미지 로드/저장
scikit-image>=0.22.0       # 이미지 비교 (SSIM)

# ============================================================
# Screen Capture
# ============================================================
mss>=9.0.1                 # 빠른 화면 캡처
pygetwindow>=0.0.9         # 윈도우 위치/크기 관리

# ============================================================
# YouTube Learning
# ============================================================
yt-dlp>=2023.11.16         # YouTube 영상 다운로드
youtube-transcript-api>=0.6.1  # YouTube 자막 추출

# ============================================================
# Video Processing
# ============================================================
# opencv-python (이미 위에 있음)
# FFmpeg는 시스템 설치 필요: winget install FFmpeg

# ============================================================
# Vision Models (선택)
# ============================================================
# openai>=1.3.0            # GPT-4V API (이미 설치됨)
# anthropic>=0.7.0         # Claude 3.5 Sonnet API (이미 설치됨)

# ============================================================
# Task Queue & Comet Integration
# ============================================================
fastapi>=0.104.1           # Task Queue Server API
uvicorn[standard]>=0.24.0  # ASGI 서버
httpx>=0.25.0              # Async HTTP client (Comet 통신)

# ============================================================
# Data Processing
# ============================================================
# 이미 설치된 것들
# pandas>=2.1.0
# numpy>=1.24.0

# ============================================================
# Logging & Monitoring
# ============================================================
# loguru>=0.7.0            # 구조화된 로깅 (이미 설치됨)

# ============================================================
# Testing & Debugging
# ============================================================
# pytest>=7.4.0            # 유닛 테스트 (이미 설치됨)
# pytest-asyncio>=0.21.0   # 비동기 테스트 (이미 설치됨)

# ============================================================
# 설치 순서
# ============================================================
# 1. Tesseract OCR 설치 (시스템):
#    winget install UB-Mannheim.TesseractOCR
#
# 2. FFmpeg 설치 (시스템):
#    winget install FFmpeg
#
# 3. Python 패키지 설치:
#    pip install -r requirements_rpa.txt
#
# 4. GPU 지원 (선택, NVIDIA GPU 있을 경우):
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
